django/requirements.txt
Django
grpcio
protobuf
psycopg2-binary
gunicorn